requirements.txt
flask==3.0.0
requests==2.31.0
chromadb==0.4.18
python-dotenv==1.0.0

